refactor : betting, winning 패키지 생성 및 수정 후 이동
