+++ src/main/java/controller/BlackjackController.java
@@ -3,10 +3,10 @@ package controller;
 import java.util.List;
 import java.util.Map;
 import model.card.Deck;
-import model.result.BettingResult;
+import model.betting.BettingResult;
 import model.participant.Dealer;
 import model.participant.Player;
-import model.result.ParticipantWinningResult;
+import model.winning.ParticipantWinningResult;
 import model.participant.Players;
 import model.turn.DealerTurn;
 import model.turn.PlayerTurn;
