+++ src/main/java/model/betting/Betting.java
@@ -1,4 +1,4 @@
-package model;
+package model.betting;
 
 import model.participant.Dealer;
