+++ src/main/java/model/winning/ParticipantWinningResult.java
@@ -1,4 +1,4 @@
-package model.result;
+package model.winning;
 
 import java.util.Collections;
 import java.util.HashMap;
