+++ src/test/java/model/BettingResultTest.java
@@ -5,14 +5,15 @@ import static org.junit.jupiter.api.Assertions.*;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import model.betting.Betting;
 import model.card.Card;
 import model.card.CardRank;
 import model.card.CardSuit;
 import model.participant.Dealer;
 import model.participant.Player;
 import model.participant.Players;
-import model.result.BettingResult;
-import model.result.ParticipantWinningResult;
+import model.betting.BettingResult;
+import model.winning.ParticipantWinningResult;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
