+++ src/test/java/model/ParticipantWinningResultTest.java
@@ -10,8 +10,8 @@ import model.card.CardSuit;
 import model.participant.Dealer;
 import model.participant.Player;
 import model.participant.Players;
-import model.result.WinningResult;
-import model.result.ParticipantWinningResult;
+import model.winning.WinningResult;
+import model.winning.ParticipantWinningResult;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
